+++ skimage/graph/__init__.py
@@ -2,7 +2,7 @@ from .spath import shortest_path
 from .mcp import MCP, MCP_Geometric, MCP_Connect, MCP_Flexible, route_through_array
 from .graph_cut import cut_threshold, cut_normalized
 from .rag import rag_mean_color, RAG, draw_rag
-from .graph_merge import merge_hierarchical, merge_hierarchical_mean_color
+from .graph_merge import merge_hierarchical
 ncut = cut_normalized
 
 
@@ -18,5 +18,4 @@ __all__ = ['shortest_path',
            'ncut',
            'draw_rag',
            'merge_hierarchical',
-           'merge_hierarchical_mean_color',
            'RAG']
